Script Changes

+++ packages/docker-mongodb/remove_mongodb_docker.sh
@@ -16,5 +16,7 @@ if [ "$(docker ps -q -f name=$MONGODB_CONTAINER_NAME)" ]; then
   docker stop $MONGODB_CONTAINER_NAME 
   echo "waiting for docker to stop"
   sleep 10s  
-  docker volume rm --force $MONGODB_VOLUME_NAME
+fi
+if [ "$(docker volume ls --format '{{.Name}}' -f name=$MONGODB_VOLUME_NAME)" ]; then
+    docker volume rm --force $MONGODB_VOLUME_NAME
 fi
